avoid Python name-mangling confusion
Another big possible source of confusion here is that Python has special
(and, I think, bizarre) rules for identifiers of the form "__Foo" (at
least two leading underscores, and at most one trailing underscore) when
found in a class definition. The identifier names get mangled to
"_ClassName__Foo" within the class, leading to possible confusion when
they're meant to refer to a global before the mangling.

This patch just changes __NodeToolCmd and __ClusterNodetoolCmd to
_NodeToolCmd and _ClusterNodetoolCmd respectively, to avoid this gotcha.