use $PAGER instead of forcing less
and exec the pager instead of spawning it in a child process, so that
when the user sends a ctrl-c or whatnot to the pager (for example, to
cancel line numbering or to cancel the "F" (follow file, like tail -f)
mode), the signal doesn't get sent to ccm and confuse things.

catching KeyboardInterrupt is a start, but we still end up with ccm
exiting before the pager does, and depending on the shell and its
configuration, the shell might try to take control of the tty at the
same time as the pager, leading to confusion.

if we really do want to Popen the pager and do things while it runs, or
afterward, we will probably want to go through the rigamarole with
setpgrp(), tcsetpgrp(), and handling SIGTTOU.