Compile Cassandra before moving to target_dir.
Previously, if compilation failed, ccm would exit with an error. But if
ccm was run again, it would see that the target directory existed, and
assume that compilation had worked.

This change defers creating the target directory until after the
compilation, so that if the compilation fails, ccm will try again the
next time it is run.