Pipe bin/cassandra stdout to devnull
On trunk the process output is now so large that it deadlocks
Popen. This was causing serious issues. According to the python
documentation, the only solution is to block with communicate()
on each call, or not to accept the output. After testing, it was
better to just pipe stdout to os.devnull and to only parse the logs
for output