Only update node configuration when necessary
Currently every call to common.make_cassandra_env() updates the node
cassandra-env file with the current install dir and node conf dir. This
method is used by many node.py commands, such as nodetool, json2sstable,
scrub, cqlsh, etc to construct the environment variables, but it's not
necessary to update the cassandra-env file every time these commands are
called, only in the first time, or when the install dir changes.

This patch makes all node commands use node.get_env() instead of
common.make_cassandra_env(), and only updates the cassandra-env file in
the first use, or when the install dir changes.

The motivation behind this is a race condition in the context of
CASSANDRA-10485, where a nodetool command is called concurrently by 2
threads, so both threads try to modify the cassandra-env file concurrently
(while it was not necessary in the first place).