Only update the envfile in __update_envfile on windows
By updating in make_cassandra_env, we were unable to guarantee
the env was updated everytime we needed it to be. Only made this
change for the windows logic though. On OSX/Linux, we're editing
cassandra.in.sh, which is in bin/ so the logic is different. On
windows we need to change cassandra-env.ps1, which needs to
be updated more often