Improve handling when not all loopback addresses are available
On Mac the user needs to make sure the loopback addresses (127.0.0.2, 127.0.0.3, etc.)
are created. Currently if they aren't created, ccm fails when creating the cluster,
but leaves it in a partial state (usually just the first node is running).

This change does a pre-flight check to make sure all loopbacks are available, print
a better warning, and exit before starting any nodes.