If the Java version is explicitly provided, use that Java version. Retrieve the distribution path from env (no matter how - it can be JAVA_HOME or JAVAx_HOME). Show a warning message if that Java version is not supported by the Cassandra version to be run.
Otherwise, use the Java version of the distribution specified in JAVA_HOME. If a Java executable is available on the PATH, its version must match the version of that distribution. In other words, CCM will fail if a Java executable is available on the PATH, but there is no JAVA_HOME defined or JAVA_HOME points to a distribution of a different version.

If no Java executable is available on the PATH and there is no JAVA_HOME defined in the env, CCM tries to guess the Java version from the Cassandra version to be run. If a JAVAn_HOME env variable points to the Java distribution with the guessed major version n, it will be set as JAVA_HOME and added to the PATH.